Qt offroad home screen (#2672)
* some cleanup first

* home sweet home

* basic home screen

* date

* fix touch

* fix mac

* hide when onroad

* box layout didn't do what i thought
